Struct webgl_rs::sampler::WebGLRSSampler

pub struct WebGLRSSampler<'ctx> { /* fields omitted */ }

Methods

impl<'ctx> WebGLRSSampler<'ctx>

pub fn delete(self)

Deletes this WebGLRSSampler object.

pub fn bind(&self, unit: u32)

Binds this WebGLRSSampler object to the texture unit at the passed index.

Arguments

unit - specifying the index of the texture unit to which to bind the sampler to.

pub fn is_valid(&self) -> bool

Returns true if this is a valid WebGLRSSampler object.

pub fn texture_compare_func(&self) -> DepthTest

Returns the texture comparison function for this WebGLRSSampler object.

pub fn set_texture_compare_func(&self, value: DepthTest)

Sets the texture comparison function for this WebGLRSSampler object.

pub fn texture_compare_mode(&self) -> CompareMode

Returns the texture comparison mode for this WebGLRSSampler object.

pub fn set_texture_compare_mode(&self, value: CompareMode)

Sets the texture comparison mode for this WebGLRSSampler object.

pub fn texture_mag_filter(&self) -> TextureMagFilter

Returns the texture magnification filter for this WebGLRSSampler object.

pub fn set_texture_mag_filter(&self, value: TextureMagFilter)

Sets the texture magnification filter for this WebGLRSSampler object.

pub fn texture_max_lod(&self) -> f32

Returns the maximum level-of-detail value for this WebGLRSSampler object.

pub fn set_texture_max_lod(&self, value: f32)

Sets the maximum level-of-detail value for this WebGLRSSampler object.

pub fn texture_min_filter(&self) -> TextureMinFilter

Returns the texture minification filter for this WebGLRSSampler object

pub fn set_texture_min_filter(&self, value: TextureMinFilter)

Sets the texture minification filter for this WebGLRSSampler object

pub fn texture_min_lod(&self) -> f32

Returns the minimum level-of-detail value for this WebGLRSSampler object.

pub fn set_texture_min_lod(&self, value: f32)

Sets the minimum level-of-detail value for this WebGLRSSampler object.

pub fn texture_wrap_r(&self) -> TextureWrap

Returns the texture wrapping function for the texture coordinate r for this WebGLRSSampler object.

pub fn set_texture_wrap_r(&self, value: TextureWrap)

Sets the texture wrapping function for the texture coordinate r for this WebGLRSSampler object.

pub fn texture_wrap_s(&self) -> TextureWrap

Return the texture wrapping function for the texture coordinate s for this WebGLRSSampler object.

pub fn set_texture_wrap_s(&self, value: TextureWrap)

Sets the texture wrapping function for the texture coordinate s for this WebGLRSSampler object.

pub fn texture_wrap_t(&self) -> TextureWrap

Returns the texture wrapping function for the texture coordinate t for this WebGLRSSampler object.

pub fn set_texture_wrap_t(&self, value: TextureWrap)

Sets the texture wrapping function for the texture coordinate t for this WebGLRSSampler object.