Struct webgl_rs::shader_program::WebGLRSShader

pub struct WebGLRSShader<'ctx> { /* fields omitted */ }

The WebGLRSShader is part of the WebGL API and can either be a vertex or a fragment shader. A WebGLRSProgram requires both types of shaders.

Methods

impl<'ctx> WebGLRSShader<'ctx>

pub fn compile(&self)

Compiles a GLSL shader into binary data so that it can be used by a WebGLRSProgram.

pub fn delete(self)

Deletes this WebGLRSShader object.

pub fn info_log(&self) -> String

Returns the information log for the specified WebGLRSShader object. It contains warnings, debugging and compile information.

pub fn shader_source(&self) -> String

Returns the source code of this WebGLRSShader as a String.

pub fn is_valid(&self) -> bool

Returns true if this WebGLRSShader is valid, false otherwise.

pub fn set_shader_source(&self, source: &str)

Sets the source code of the WebGLRSShader object.

pub fn delete_status(&self) -> bool

Returns a bool indicating whether or not the WebGLRSShader is flagged for deletion.

pub fn compile_status(&self) -> bool

Returns a bool indicating whether or not the last compilation was successful.

pub fn kind(&self) -> ShaderKind

Returns a ShaderKind value indicating whether the shader is a vertex shader or fragment shader.