Struct webgl_rs::shader_program::WebGLRSProgram

pub struct WebGLRSProgram<'ctx> { /* fields omitted */ }

The WebGLRSProgram is part of the WebGL API and is a combination of two compiled WebGLShaders consisting of a vertex shader and a fragment shader (both written in GLSL). These are then linked into a usable program.

Methods

impl<'ctx> WebGLRSProgram<'ctx>

pub fn delete(self)

Deletes this WebGLRSProgram object. This method has no effect if the program has already been deleted.

pub fn detach_shader(&self, shader: &WebGLRSShader)

Detaches a previously attached WebGLRSShader from this WebGLRSProgram.

Arguments

shader - shader to detach

pub fn attach_shader(&self, shader: &WebGLRSShader)

Attaches a WebGLRSShader to this WebGLRSProgram.

Arguments

shader - shader to attach

pub fn bind_attrib_location(&self, index: u32, name: &str)

Binds a generic vertex index to an attribute variable for this WebGLRSProgram

Arguments

index - specifying the index of the generic vertex to bind.

name - specifying the name of the variable to bind to the generic vertex index. This name cannot start with "webgl_" or "webgl", as these are reserved for use by WebGL.

pub fn info_log(&self) -> String

Returns the information log for this WebGLRSProgram object. It contains errors that occurred during failed linking or validation of this object.

pub fn is_valid(&self) -> bool

Returns true if this WebGLRSProgram is valid, false otherwise.

pub fn link(&self)

Links this WebGLRSProgram to the attached vertex and fragment shaders.

pub fn enable(&self)

Sets this WebGLRSProgram as part of the current rendering state.

pub fn validate(&self)

Validates this WebGLRSProgram. Checks if it is successfully linked and if it can be used in the current WebGL state.

pub fn active_attrib(&self, index: u32) -> WebGLActiveInfo

Returns a WebGLActiveInfo object containing size, type, and name of a vertex attribute.

It is generally used when querying unknown uniforms either for debugging or generic library creation.

Arguments

index - specifying the index of the vertex attribute to get.

pub fn active_uniform(&self, index: u32) -> WebGLActiveInfo

Returns a WebGLActiveInfo object containing size, type, and name of a uniform attribute.

It is generally used when querying unknown uniforms either for debugging or generic library creation.

Arguments

index - specifying the index of the vertex attribute to get.

pub fn attrib_location(&self, name: &str) -> u32

Returns the location of an attribute variable in this WebGLRSProgram.

Arguments

name - specifying the name of the attribute variable whose location to get.

pub fn uniform_location(&self, name: &str) -> WebGLRSUniformLocation

Returns the location of a specific uniform variable which is part this WebGLRSProgram.

The uniform variable is returned as a WebGLUniformLocation object, which is an opaque identifier used to specify where in the GPU's memory that uniform variable is located.

Arguments

name- specifying the name of the uniform variable whose location is to be returned. The name can't have any whitespace in it, and you can't use this function to get the location of any uniforms starting with the reserved string "gl_", since those are internal to the WebGL layer.

pub fn frag_data_location(&self, name: &str) -> i32

Returns the binding of color numbers to user-defined varying out variables.

Arguments

name - specifying the name of the user-defined varying out variable.

pub fn transform_feedback_varying(&self, index: u32) -> WebGLActiveInfo

Returns information about varying variables from WebGLTransformFeedback buffers.

Arguments

index - specifying the index of the varying variable whose information to retrieve.

pub fn uniform_block_index(&self, uniform_block_name: &str) -> u32

Retrieves the index of a uniform block in this WebGLRSProgram.

Arguments

uniform_block_name - specifying the name of the uniform block to whose index to retrieve.

pub fn active_uniform_block_name(&self, uniform_block_index: u32) -> String

Retrieves the name of the active uniform block at a given index in this WebGLRSProgram.

Arguments

uniform_block_index - specifying the index of the uniform block to whose name to retrieve.

pub fn assign_uniform_block_binding(
&self,
uniform_block_index: u32,
uniform_block_binding: u32
)

Asigns binding points for active uniform blocks.

Arguments

uniform_block_index - specifying the index of the active uniform block within the program.

uniform_block_binding - specifying the binding point to which to bind the uniform block.

pub fn delete_status(&self) -> bool

Indicates whether or not the program is flagged for deletion.

pub fn link_status(&self) -> bool

Indicates whether or not the last link operation was successful.

pub fn validate_status(&self) -> bool

Indicates whether or not the last validation operation was successful.

pub fn attached_shaders(&self) -> i32

Returns the number of attached shaders to a program.

pub fn active_attributes(&self) -> i32

Returns the number of active attribute variables to a program.

pub fn active_uniforms(&self) -> i32

Returns the number of active uniform variables to a program.

pub fn transform_feedback_buffer_mode(&self) -> TransformFeedbackBufferMode

Returns the buffer mode when transform feedback is active. May be SeparateAttribs or InterleavedAttribs.

pub fn transform_feedback_varyings(&self) -> i32

Returns the number of varying variables to capture in transform feedback mode.

pub fn active_uniform_blocks(&self) -> i32